backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Configuration and environment
python-dotenv>=1.0.0
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1

# Graph database
neo4j>=5.16.0

# Vector database
pinecone>=5.0.0

# Cache
redis>=5.0.1

# AI/ML APIs
anthropic>=0.18.1
openai>=1.10.0
tiktoken>=0.5.0

# Numerical computing (for vector operations)
numpy>=1.24.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Encryption for OAuth tokens
cryptography>=42.0.1

# HTTP client
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.2
aiofiles>=23.2.1

# Document processing
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
markdown>=3.5.0
lxml>=5.1.0

# Rate limiting
slowapi>=0.1.9

# Text processing
langdetect>=1.0.9

# Retry logic
tenacity>=8.2.0

# Monitoring
prometheus-client>=0.19.0

# Development tools
ruff>=0.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Error tracking
sentry-sdk[fastapi]>=1.50.0
